+++ lib/tracer.js
